@@ -1,168 +1,203 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:flutter_application/views/widgets/base_screen.dart';
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'dart:convert';
+import 'package:shared_preferences/shared_preferences.dart';
+
+const Color myColor = Color(0xFF9D4EDD);
 
 class DonationsScreen extends StatefulWidget {
-  const DonationsScreen({Key? key}) : super(key: key);
+  const DonationsScreen({super.key});
 
   @override
   DonationsScreenState createState() => DonationsScreenState();
 }
 
 class DonationsScreenState extends State<DonationsScreen> {
-  int _page = 1; // Índice actual para la barra de navegación inferior
+  int _page = 1;
+  String? donationType;
+  String title = '';
+  String description = '';
+  final TextEditingController _titleController = TextEditingController();
+  final TextEditingController _descriptionController = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          'Donaciones',
-          style: TextStyle(
-            color: Colors.black,
-            fontSize: 24,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        elevation: 0,
-        centerTitle: true,
-        // Se han removido los íconos de notificaciones y perfil
-      ),
+    donationType = ModalRoute.of(context)?.settings.arguments as String?;
+    return BaseScreen(
+      title: 'Donaciones - ${donationType ?? "No argumento"}',
+      appBarBackgroundColor: Colors.white,
+      appBarElevation: 10,
+      appBarShadowColor: Colors.black.withOpacity(0.5),
+      appBarSurfaceTintColor: Colors.transparent,
+      appBarIconTheme: const IconThemeData(color: Colors.black),
+      currentPage: _page,
+      showAppBarActions: false,
+      onBottomNavTap: (index) {
+        if (index == 0) {
+          Navigator.pushNamed(context, '/');
+        } else if (index == 1) {
+          Navigator.pushNamed(context, '/object');
+        } else if (index == 2) {
+          Navigator.pushNamed(context, '/map');
+        }
+        setState(() {
+          _page = index;
+        });
+      },
       body: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            titleInput(TextEditingController()),
+            titleInput(_titleController),
             const SizedBox(height: 24),
-            descriptionInput(TextEditingController()),
+            descriptionInput(_descriptionController),
             const SizedBox(height: 24),
             buildButtonContinue(),
           ],
         ),
       ),
-      bottomNavigationBar: CurvedNavigationBar(
-        index: _page,
-        height: 60.0,
-        items: const <Widget>[
-          Icon(Icons.home, size: 30, color: Colors.black),
-          Icon(Icons.add, size: 30, color: Colors.black),
-          Icon(Icons.map_outlined, size: 30, color: Colors.black),
-        ],
-        color: const Color(0xFFDEC3BE),
-        buttonBackgroundColor: const Color(0xFFDEC3BE),
-        backgroundColor: Colors.white,
-        animationCurve: Curves.easeInOut,
-        animationDuration: const Duration(milliseconds: 300),
-        onTap: (index) {
-          if (index == 0) {
-            Navigator.pushNamed(context, '/');
-          } else if (index == 1) {
-            Navigator.pushNamed(context, '/object');
-          } else if (index == 2) {
-            Navigator.pushNamed(context, '/map');
-          }
-          setState(() {
-            _page = index;
-          });
-        },
-      ),
     );
   }
-}
 
-Widget titleInput(TextEditingController controller) {
-  return Column(
-    crossAxisAlignment: CrossAxisAlignment.start,
-    children: [
-      const Text(
-        'Título',
-        style: TextStyle(
-          fontSize: 22,
-          fontWeight: FontWeight.bold,
+  void saveData() async {
+    if (title.isEmpty || description.isEmpty) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text('Por favor, ingresa todos los campos')),
+      );
+      return;
+    }
+
+    User? user = FirebaseAuth.instance.currentUser;
+    if (user == null) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text('Por favor, inicia sesión')),
+      );
+      return;
+    }
+
+    String userId = user.uid;
+    showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (context) => const AlertDialog(
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            CircularProgressIndicator(),
+            SizedBox(height: 16),
+            Text('Guardando donación...')
+          ],
         ),
       ),
-      const SizedBox(height: 10),
-      Container(
-        width: double.infinity,
-        height: 60,
-        decoration: BoxDecoration(
-          color: const Color(0xFFDEC3BE),
-          borderRadius: BorderRadius.circular(16),
+    );
+
+    try {
+      var docRef = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(userId)
+          .collection('donations')
+          .add({
+        "userId": userId,
+        "type": donationType ?? "No type",
+        "title": title,
+        "description": description,
+        "estado": 0,
+        "createdAt": DateTime.now().toIso8601String(),
+      });
+
+      String donationId = docRef.id;
+      Map<String, dynamic> donationData = {
+        "donationId": donationId,
+        "userId": userId,
+        "type": donationType ?? "No type",
+        "title": title,
+        "description": description,
+        "estado": 0,
+        "createdAt": DateTime.now().toIso8601String(),
+      };
+
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      await prefs.setString('lastDonation', jsonEncode(donationData));
+
+      Navigator.pop(context);
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text('Donación guardada con éxito')),
+      );
+
+      Future.delayed(const Duration(milliseconds: 500), () {
+        Navigator.pushNamed(context, '/map');
+      });
+    } catch (e) {
+      Navigator.pop(context);
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text('Error al guardar la donación: $e')),
+      );
+    }
+  }
+
+  Widget titleInput(TextEditingController controller) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          'Título',
+          style: GoogleFonts.amaticSc(fontSize: 22),
         ),
-        child: TextField(
+        const SizedBox(height: 10),
+        TextField(
           controller: controller,
-          decoration: InputDecoration(
+          onChanged: (value) => setState(() => title = value),
+          decoration: const InputDecoration(
             hintText: 'Escribe el título aquí',
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(16),
-              borderSide: BorderSide.none,
-            ),
-            contentPadding:
-                const EdgeInsets.symmetric(horizontal: 20, vertical: 18),
-            filled: true,
-            fillColor: const Color(0xFFDEC3BE),
           ),
         ),
-      ),
-    ],
-  );
-}
+      ],
+    );
+  }
 
-Widget descriptionInput(TextEditingController controller) {
-  return Column(
-    crossAxisAlignment: CrossAxisAlignment.start,
-    children: [
-      const Text(
-        'Descripción',
-        style: TextStyle(
-          fontSize: 22,
-          fontWeight: FontWeight.bold,
+  Widget descriptionInput(TextEditingController controller) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          'Descripción',
+          style: GoogleFonts.amaticSc(fontSize: 22),
         ),
-      ),
-      const SizedBox(height: 10),
-      Container(
-        padding: const EdgeInsets.all(16),
-        height: 150,
-        decoration: BoxDecoration(
-          color: Colors.grey[200],
-          borderRadius: BorderRadius.circular(20),
-          border: Border.all(color: Colors.grey),
-        ),
-        child: TextField(
+        const SizedBox(height: 10),
+        TextField(
           controller: controller,
-          maxLines: null,
-          expands: true,
-          decoration: const InputDecoration.collapsed(
+          onChanged: (value) => setState(() => description = value),
+          maxLines: 5,
+          decoration: const InputDecoration(
             hintText: 'Agrega una descripción detallada aquí...',
           ),
         ),
-      ),
-    ],
-  );
-}
+      ],
+    );
+  }
 
-Widget buildButtonContinue() {
-  return Builder(
-    builder: (context) => ElevatedButton(
-      onPressed: () {
-        print("Botón Continuar presionado");
-        Navigator.pushNamed(context, '/map');
-      },
+  Widget buildButtonContinue() {
+    return ElevatedButton(
+      onPressed: saveData,
       style: ElevatedButton.styleFrom(
-        backgroundColor: const Color(0xFFDEC3BE),
+        backgroundColor: myColor,
         minimumSize: const Size(150, 50),
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(20),
         ),
       ),
-      child: const Text(
+      child: Text(
         'Continuar',
-        style: TextStyle(
+        style: GoogleFonts.poppins(
           color: Colors.white,
           fontSize: 20,
         ),
       ),
-    ),
-  );
+    );
+  }
 }
