@@ -1,27 +1,87 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_application/views/widgets/base_screen.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+const Color myColor = Color(0xFF9D4EDD);
+const double imageSize = 75.0;
 
 class ObjectScreen extends StatefulWidget {
-  const ObjectScreen({Key? key}) : super(key: key);
+  const ObjectScreen({super.key});
 
   @override
   ObjectScreenState createState() => ObjectScreenState();
 }
 
 class ObjectScreenState extends State<ObjectScreen> {
-  late Color myColor;
   late Size mediaSize;
-  int _page = 1; // Índice inicial para la barra inferior
+  int _page = 1;
 
   @override
   Widget build(BuildContext context) {
     mediaSize = MediaQuery.of(context).size;
-    myColor = const Color(0xFFDEC3BE);
+
+    Widget bodyWithGradient = Container(
+      decoration: const BoxDecoration(
+        gradient: LinearGradient(
+          colors: [
+            Color(0xFFEDE7F6),
+            Color(0xFFF3E5F5),
+          ],
+          begin: Alignment.topCenter,
+          end: Alignment.bottomCenter,
+        ),
+      ),
+      child: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                _buildCategoryButton(
+                  label: 'Comida',
+                  icon: Icons.fastfood,
+                  donationArgument: 'Comida',
+                ),
+                const SizedBox(width: 20),
+                _buildCategoryButton(
+                  label: 'Ropa',
+                  icon: Icons.checkroom,
+                  donationArgument: 'Ropa',
+                ),
+              ],
+            ),
+            const SizedBox(height: 20),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                _buildCategoryButton(
+                  label: 'Juguetes',
+                  icon: Icons.toys,
+                  donationArgument: 'Juguetes',
+                ),
+                const SizedBox(width: 20),
+                _buildCategoryButton(
+                  label: 'Mascotas',
+                  icon: Icons.pets,
+                  donationArgument: 'Mascotas',
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
 
     return BaseScreen(
-      title: 'Object', // Título del AppBar
+      title: 'Elige el objeto a donar',
+      appBarBackgroundColor: Colors.white,
+      appBarElevation: 10,
+      appBarShadowColor: Colors.black.withOpacity(0.5),
+      appBarSurfaceTintColor: Colors.transparent,
+      appBarIconTheme: const IconThemeData(color: Colors.black),
       currentPage: _page,
-      showAppBarActions: false, // Oculta íconos de notificaciones y perfil
+      showAppBarActions: false,
       onBottomNavTap: (index) {
         if (index == 0) {
           Navigator.pushNamed(context, '/');
@@ -34,110 +94,50 @@ class ObjectScreenState extends State<ObjectScreen> {
           _page = index;
         });
       },
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                _buildBottomFood(mediaSize),
-                const SizedBox(width: 20),
-                _buildBottomClothes(mediaSize),
-              ],
+      body: bodyWithGradient,
+    );
+  }
+
+  Widget _buildCategoryButton({
+    required String label,
+    required IconData icon,
+    required String donationArgument,
+  }) {
+    Widget buttonContent;
+    if (donationArgument == 'Comida') {
+      buttonContent = Image.asset('assets/fonts/imagen_comida.png', width: imageSize, height: imageSize);
+    } else if (donationArgument == 'Mascotas') {
+      buttonContent = Image.asset('assets/fonts/imagen_animal.png', width: imageSize, height: imageSize);
+    } else if (donationArgument == 'Juguetes') {
+      buttonContent = Image.asset('assets/fonts/imagen_juego.png', width: imageSize, height: imageSize);
+    } else if (donationArgument == 'Ropa') {
+      buttonContent = Image.asset('assets/fonts/imagen_ropa.png', width: imageSize, height: imageSize);
+    } else {
+      buttonContent = Icon(icon, color: myColor, size: 30);
+    }
+
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        ElevatedButton(
+          onPressed: () {
+            Navigator.pushNamed(context, '/donations', arguments: donationArgument);
+          },
+          style: ElevatedButton.styleFrom(
+            backgroundColor: Colors.white,
+            minimumSize: const Size(150, 150),
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
             ),
-            const SizedBox(height: 20),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                _buildBottomToys(mediaSize),
-                const SizedBox(width: 20),
-                _buildBottomPets(mediaSize),
-              ],
-            ),
-          ],
+          ),
+          child: buttonContent,
         ),
-      ),
-    );
-  }
-
-  Widget _buildBottomFood(Size mediaSize) {
-    return ElevatedButton(
-      onPressed: () {
-        Navigator.pushNamed(context, '/donations');
-      },
-      style: ElevatedButton.styleFrom(
-        backgroundColor: myColor,
-        minimumSize: const Size(150, 150),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
+        const SizedBox(height: 8),
+        Text(
+          label,
+          style: GoogleFonts.poppins(fontSize: 16, fontWeight: FontWeight.normal, color: Colors.black),
         ),
-      ),
-      child: const Icon(
-        Icons.fastfood,
-        color: Colors.white,
-        size: 30,
-      ),
-    );
-  }
-
-  Widget _buildBottomClothes(Size mediaSize) {
-    return ElevatedButton(
-      onPressed: () {
-        Navigator.pushNamed(context, '/donations');
-      },
-      style: ElevatedButton.styleFrom(
-        backgroundColor: myColor,
-        minimumSize: const Size(150, 150),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-      ),
-      child: const Icon(
-        Icons.add,
-        color: Colors.white,
-        size: 30,
-      ),
-    );
-  }
-
-  Widget _buildBottomToys(Size mediaSize) {
-    return ElevatedButton(
-      onPressed: () {
-        Navigator.pushNamed(context, '/donations');
-      },
-      style: ElevatedButton.styleFrom(
-        backgroundColor: myColor,
-        minimumSize: const Size(150, 150),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-      ),
-      child: const Icon(
-        Icons.toys,
-        color: Colors.white,
-        size: 30,
-      ),
-    );
-  }
-
-  Widget _buildBottomPets(Size mediaSize) {
-    return ElevatedButton(
-      onPressed: () {
-        Navigator.pushNamed(context, '/donations');
-      },
-      style: ElevatedButton.styleFrom(
-        backgroundColor: myColor,
-        minimumSize: const Size(150, 150),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-      ),
-      child: const Icon(
-        Icons.pets,
-        color: Colors.white,
-        size: 30,
-      ),
+      ],
     );
   }
 }
