@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:google_fonts/google_fonts.dart';
 
-const Color myColor = Color(0xFFDEC3BE);
+const Color myColor = Color(0xFF9D4EDD);
 
 class ProfileScreen extends StatefulWidget {
   const ProfileScreen({Key? key}) : super(key: key);
@@ -10,105 +12,98 @@ class ProfileScreen extends StatefulWidget {
 }
 
 class _ProfileScreenState extends State<ProfileScreen> {
-  // Si se obtiene la foto del usuario (por ejemplo, de Google), se asigna a esta variable;
-  // de lo contrario, se usa una imagen por defecto.
   String profileImageUrl = '';
-  final TextEditingController _nameController =
-      TextEditingController(text: 'Juan Pérez');
+  final TextEditingController _nameController = TextEditingController();
   final TextEditingController _phoneController = TextEditingController();
-  int donationsCount = 5; // Ejemplo de cantidad de donaciones
+  int donationsCount = 5;
+
+  @override
+  void initState() {
+    super.initState();
+    _loadUserData();
+  }
+
+  void _loadUserData() {
+    User? user = FirebaseAuth.instance.currentUser;
+    if (user != null) {
+      setState(() {
+        profileImageUrl = user.photoURL ?? '';
+        _nameController.text = user.displayName ?? 'Juan Pérez';
+      });
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text(
+        title: Text(
           'Perfil',
-          style: TextStyle(
+          style: GoogleFonts.amaticSc(
             color: Colors.black,
             fontSize: 24,
-            fontWeight: FontWeight.bold,
+            fontWeight: FontWeight.normal,
           ),
         ),
-        backgroundColor: Colors.white,
-        elevation: 0,
         centerTitle: true,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.settings, color: Colors.black),
-            onPressed: () {
-              print('Configuración');
-              Navigator.pushNamed(context, '/configuration');
-            },
-          ),
-        ],
+        backgroundColor: Colors.white,
+        elevation: 10,
+        shadowColor: Colors.black.withOpacity(0.5),
+        surfaceTintColor: Colors.transparent,
+        iconTheme: const IconThemeData(color: Colors.black),
       ),
-      backgroundColor: Colors.grey[200],
+      backgroundColor: Colors.white,
       body: SingleChildScrollView(
         padding: const EdgeInsets.all(16.0),
         child: Column(
           children: [
-            // La foto de perfil permanece fuera del Card.
+            const SizedBox(height: 20),
             _buildProfilePicture(),
-            const SizedBox(height: 24),
-            // La tarjeta de donaciones también queda por fuera.
+            const SizedBox(height: 20),
             _buildDonationCard(),
-            const SizedBox(height: 24),
-            // Se agrupan los campos editables y el botón dentro de un Card.
-            Card(
-              elevation: 8,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(16),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Column(
-                  children: [
-                    _buildNameField(),
-                    const SizedBox(height: 16),
-                    _buildPhoneField(),
-                    const SizedBox(height: 40),
-                    _buildSaveButton(),
-                  ],
-                ),
-              ),
-            ),
+            const SizedBox(height: 20),
+            _buildInfoCard(),
+            const SizedBox(height: 40),
+            _buildSaveButton(),
           ],
         ),
       ),
     );
   }
 
-  // Widget para la foto de perfil.
   Widget _buildProfilePicture() {
     return CircleAvatar(
-      radius: 50,
-      backgroundColor: myColor,
+      radius: 70,
+      backgroundColor: Colors.white,
       child: profileImageUrl.isEmpty
-          ? const Icon(Icons.person, size: 50, color: Colors.white)
-          : ClipOval(child: Image.network(profileImageUrl, fit: BoxFit.cover)),
+          ? Icon(Icons.person, size: 70, color: myColor)
+          : ClipOval(
+              child: Image.network(
+                profileImageUrl,
+                fit: BoxFit.cover,
+                width: 140,
+                height: 140,
+              ),
+            ),
     );
   }
 
-  // Widget para la tarjeta de donaciones.
   Widget _buildDonationCard() {
     return Card(
       color: myColor,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-      ),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
       child: Padding(
         padding: const EdgeInsets.symmetric(vertical: 16.0, horizontal: 24.0),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            const Text(
+            Text(
               'Donaciones',
-              style: TextStyle(color: Colors.white, fontSize: 18),
+              style: GoogleFonts.amaticSc(color: Colors.white, fontSize: 20),
             ),
             Text(
               '$donationsCount',
-              style: const TextStyle(color: Colors.white, fontSize: 18),
+              style: GoogleFonts.amaticSc(color: Colors.white, fontSize: 20),
             ),
           ],
         ),
@@ -116,40 +111,63 @@ class _ProfileScreenState extends State<ProfileScreen> {
     );
   }
 
-  // Widget para el campo del nombre.
-  Widget _buildNameField() {
-    return TextFormField(
-      controller: _nameController,
-      decoration: const InputDecoration(
-        labelText: 'Nombre',
-        border: OutlineInputBorder(),
+  Widget _buildInfoCard() {
+    return Card(
+      elevation: 4,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      child: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 24.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            _buildReadOnlyField(label: 'Nombre', controller: _nameController),
+            const SizedBox(height: 16),
+            _buildReadOnlyField(label: 'Teléfono', controller: _phoneController),
+          ],
+        ),
       ),
     );
   }
 
-  // Widget para el campo del teléfono.
-  Widget _buildPhoneField() {
-    return TextFormField(
-      controller: _phoneController,
-      decoration: const InputDecoration(
-        labelText: 'Teléfono',
-        border: OutlineInputBorder(),
-      ),
+  Widget _buildReadOnlyField({required String label, required TextEditingController controller}) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          label,
+          style: GoogleFonts.amaticSc(fontSize: 20, fontWeight: FontWeight.normal),
+        ),
+        const SizedBox(height: 4),
+        TextFormField(
+          controller: controller,
+          readOnly: true,
+          style: GoogleFonts.poppins(fontWeight: FontWeight.normal),
+          decoration: InputDecoration(
+            filled: true,
+            fillColor: Colors.grey[100],
+            border: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(12),
+              borderSide: const BorderSide(color: Colors.transparent),
+            ),
+          ),
+        ),
+      ],
     );
   }
 
-  // Widget para el botón de guardar cambios.
   Widget _buildSaveButton() {
     return ElevatedButton(
       onPressed: () {
-        // Aquí puedes agregar la lógica para guardar los cambios.
         print('Guardar cambios');
       },
       style: ElevatedButton.styleFrom(
         backgroundColor: myColor,
         minimumSize: const Size.fromHeight(50),
       ),
-      child: const Text('Guardar cambios'),
+      child: Text(
+        'Guardar cambios',
+        style: GoogleFonts.poppins(color: Colors.white, fontWeight: FontWeight.bold),
+      ),
     );
   }
 }
